Add files via upload
Water plays a vital role in everyone's life, 70% of the human body is made of water. Contaminated water and poor sanitation are linked to the transmission of diseases such as cholera, diarrhoea, dysentery, hepatitis A, typhoid, and polio. A superior quality of water is crucial to the economic, health, and social well-being of the people. Monitoring the quality of your water and testing it regularly is very important to maintain reliable and safe water sources and eliminate the potential health risks related to water contamination.
Our web application, HydroSafe, informs users whether a given sample of water is safe to drink or not based on some specifications provided. The prediction of water quality is done by a machine learning model that uses Random forest classifier algorithm. If the water is not potable, our application also provides suggestions on the alternative uses of this water sample, thus providing quick water analysis along with better usage of both potable and non-potable forms of water.
Project Demo:
Drive link: https://drive.google.com/file/d/1otX9eHJrQtjZP1B61WMKoYkYc9-oclmK/view?usp=sharing